scripts/requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
tweepy>=4.10.0
pymongo>=4.2.0
pandas>=1.4.0
numpy>=1.22.0
beautifulsoup4>=4.11.0
